Removed compiler warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@11024 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/modes/linear_world.cpp
@@ -353,7 +353,7 @@ void LinearWorld::newLap(unsigned int kart_index)
  */
 int LinearWorld::getSectorForKart(const AbstractKart *kart) const
 {
-    if(kart->getWorldKartId()>=(int)m_kart_info.size())
+    if(kart->getWorldKartId()>=m_kart_info.size())
         return QuadGraph::UNKNOWN_SECTOR;
     return m_kart_info[kart->getWorldKartId()].getSector()
           ->getCurrentGraphNode();
